requirements.txt
azure-functions
requests
flair
torch